Refactor the network to allow for external parameters
I have refactored all of the network classes to allow for
'external system parameters' to be provided by the user. These
parameters can for example capture whether the system is switched
on, or what mode it is -- since these characteristics are
essential in determining the observed state of the system.